refactor: Split SliceHeaderControls into smaller files (#31270)
